more helium tweaks

+++ include/constants.h
@@ -55,6 +55,6 @@ extern const int prev_pin;
 //time checks
 extern unsigned long previousMillis;
 extern unsigned long currentMillis;
-extern unsigned long deltaMillis;
+extern double deltaMillis;
 
 #endif
